fix test server startup; refresh token missing fields;

+++ go/integrations/integration_test.go
@@ -16,11 +16,13 @@ func TestMain(m *testing.M) {
 	if err != nil {
 		panic(err)
 	}
-	defer func() {
-		if err := cmd.Process.Kill(); err != nil {
-			fmt.Printf("Failed to close server: %v", util.ErrCheck(err))
-		}
-	}()
+	if cmd != nil {
+		defer func() {
+			if err := cmd.Process.Kill(); err != nil {
+				fmt.Printf("Failed to close server: %v", util.ErrCheck(err))
+			}
+		}()
+	}
 
 	code := m.Run()
